chnaged time zone name to long

+++ src/course-home/dates-tab/timeline/Day.jsx
@@ -78,7 +78,7 @@ const Day = ({
                       <span className="mx-1">due</span>
                       <FormattedTime
                         value={item.date}
-                        timeZoneName="short"
+                        timeZoneName="long"
                         {...timezoneFormatArgs}
                       />
                     </span>
